fixed css stuff and removed useless package
turns out that my huge commit yesterday was made useless by the fact that in the process of suffering through SVG loaders, I made a method to replace colors in an image. So... I don't need svgs now!

Made CSS more organized; one file (the theme folders are still there but we can take that out)